requirements/local.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncio==4.0.0
attrs==25.3.0
boolean.py==5.0
CacheControl==0.14.3
certifi==2025.7.9
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssbeautifier==1.15.4
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
distlib==0.3.9
Django==5.2.3
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-extensions==4.1
django-redis==6.0.0
django-stubs==5.2.2
django-stubs-ext==5.2.2
djlint==1.36.4
EditorConfig==0.17.1
fast-depends==2.4.12
faststream==0.5.43
filelock==3.18.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
jsbeautifier==1.15.4
json5==0.12.0
license-expression==30.4.3
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.3
mypy==1.17.0
mypy_extensions==1.1.0
nodeenv==1.9.1
orjson==3.10.18
packageurl-python==0.17.1
packaging==25.0
pathspec==0.12.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
py-serializable==2.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruff==0.12.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
starlette==0.47.1
structlog==25.4.0
toml==0.10.2
tqdm==4.67.1
types-PyYAML==6.0.12.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
